Desktop/Project/AI-Powered-SQL-Query-Generator-main/requirements.txt
# AI Frameworks
langchain
llama-index
crewai
chromadb

# LLMs & Embeddings
openai
google-generativeai
transformers
sentence-transformers

# Backend & API
fastapi
uvicorn
pydantic

# Database Drivers
sqlalchemy
psycopg2-binary   # PostgreSQL
pyodbc        # SQL Server
mysql-connector-python
snowflake-connector-python

# Additional Tools
tqdm
python-dotenv      # Environment Variables
loguru            # Logging

# Testing
pytest

#sqlparse
sqlparse

streamlit   